refactor(components): replace and remove old link components
