+++ src/components/History/NetworkUpgradeSummary.tsx
@@ -10,7 +10,7 @@ import { getLocaleForNumberFormat } from "@/lib/utils/translations"
 
 import NetworkUpgradeSummaryData from "../../data/NetworkUpgradeSummaryData"
 import Emoji from "../Emoji"
-import InlineLink from "../Link"
+import InlineLink from "../ui/Link"
 
 type NetworkUpgradeSummaryProps = {
   name: string
