+++ src/components/LanguagePicker/NoResultsCallout.tsx
@@ -1,8 +1,6 @@
 import { useTranslation } from "next-i18next"
 
-import { BaseLink } from "@/components/Link"
-
-import MenuItem from "./MenuItem"
+import { BaseLink } from "../ui/Link"
 
 type NoResultsCalloutProps = { onClose: () => void }
 
@@ -14,9 +12,7 @@ const NoResultsCallout = ({ onClose }: NoResultsCalloutProps) => {
       <p className="text-body-medium">
         {t("page-languages-want-more-paragraph")}
       </p>
-      {/* TODO: use migrated Link component */}
       <BaseLink
-        as={MenuItem}
         key="item-no-results"
         href="contributing/translation-program"
         onClick={onClose}
