+++ src/components/LanguagePicker/index.tsx
@@ -1,8 +1,6 @@
 import { useRouter } from "next/router"
 import { useTranslation } from "next-i18next"
 
-import { BaseLink } from "@/components/Link"
-
 import { cn } from "@/lib/utils/cn"
 
 import {
@@ -13,6 +11,7 @@ import {
   CommandList,
 } from "../ui/command"
 import { Dialog, DialogContent, DialogTrigger } from "../ui/dialog"
+import { BaseLink } from "../ui/Link"
 import { Popover, PopoverContent, PopoverTrigger } from "../ui/popover"
 
 import MenuItem from "./MenuItem"
