+++ src/components/Layer2NetworksTable/hooks/useNetworkColumns.tsx
@@ -9,9 +9,9 @@ import { ExtendedRollup, Lang } from "@/lib/types"
 import { TableMeta } from "@/components/DataTable"
 import { Image } from "@/components/Image"
 import NetworkMaturityTooltip from "@/components/Layer2NetworksTable/NetworkMaturityTooltip"
-import InlineLink from "@/components/Link"
 import Tooltip from "@/components/Tooltip"
 import { Button } from "@/components/ui/buttons/Button"
+import InlineLink from "@/components/ui/Link"
 import { TableCell, TableHead } from "@/components/ui/table"
 
 import { cn } from "@/lib/utils/cn"
