+++ src/pages/404.tsx
@@ -3,9 +3,9 @@ import { serverSideTranslations } from "next-i18next/serverSideTranslations"
 
 import { BasePageProps, Lang } from "@/lib/types"
 
-import InlineLink from "@/components/Link"
 import MainArticle from "@/components/MainArticle"
 import Translation from "@/components/Translation"
+import InlineLink from "@/components/ui/Link"
 
 import { existsNamespace } from "@/lib/utils/existsNamespace"
 import { getLastDeployDate } from "@/lib/utils/getLastDeployDate"
